Time: 769 ms (30.79%), Space: 58 MB (50.12%) - LeetHub

+++ 3851-find-sum-of-array-product-of-magical-sequences/3851-find-sum-of-array-product-of-magical-sequences.cpp
@@ -97,7 +97,7 @@ public:
         int mod=1e9+7;
         vector<ll>fact;
         vector<ll>inverseFact;
-        map<tuple<int,int,int,int>,int>dp;
+        map<tuple<int,int,int,int>,ll>dp;
         ll findPower(ll a,ll b){
             // if(b==0) return 1;
             // ll ans=1;
